requirements.txt
numpy
openai
Pillow
python-dotenv
Requests
streamlit
tensorflow
